requirements.txt
accelerate==1.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
asyncer==0.0.7
attrs==25.1.0
avidtools==0.1.2
backoff==2.2.1
base2048==0.1.3
bcrypt==4.2.1
beautifulsoup4==4.13.3
bert-score==0.3.13
bidict==0.23.1
blinker==1.9.0
blis==1.2.0
bokeh==3.4.3
boto3==1.36.26
botocore==1.36.26
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chainlit==2.1.2
chardet==5.2.0
charset-normalizer==3.4.1
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
cmd2==2.4.3
cohere==4.57
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
databricks-sdk==0.44.1
dataclasses-json==0.6.7
datasets==2.16.1
DateTime==5.5
decorator==5.1.1
deepl==1.17.0
Deprecated==1.2.18
dill==0.3.7
distro==1.9.0
docopt==0.6.2
durationpy==0.9
ecoji==0.1.1
elastic-transport==8.17.0
elasticsearch==8.17.1
emoji==2.14.1
evaluate==0.4.3
executing==2.2.0
faiss-cpu==1.8.0
fastapi==0.115.8
fastavro==1.10.0
fastembed==0.5.1
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
fonttools==4.56.0
frozenlist==1.5.0
fschat==0.2.36
fsspec==2023.10.0
fuzzywuzzy==0.18.0
garak==0.10.2
giskard==2.16.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
griffe==0.48.0
groq==0.18.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
ipdb==0.13.13
ipython==8.32.0
it_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/it_core_news_lg-3.8.0/it_core_news_lg-3.8.0-py3-none-any.whl#sha256=b78582d0b2d05fd6509995f68ab7452efed7a27c6fbc5a071e9a9787a58c1e87
it_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-3.8.0/it_core_news_sm-3.8.0-py3-none-any.whl#sha256=3f617bf9a8ae0418953cf1fbf014e10272684c4229e882a7fd748b637d0100bf
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.30.3
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==32.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-neo4j==0.3.0
langchain-ollama==0.2.3
langchain-text-splitters==0.3.6
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.3.8
language_data==1.3.0
latex2mathml==3.77.0
Lazify==0.4.0
Levenshtein==0.26.1
litellm==1.50.4
literalai==0.1.103
llvmlite==0.44.0
loguru==0.7.3
lorem==0.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mixpanel==4.10.1
mlflow-skinny==2.20.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.25.2
nemollm==0.3.5
neo4j==5.28.1
neo4j-graphrag==1.4.3
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.20
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
nvdlib==0.7.9
oauthlib==3.2.2
octoai-sdk==0.10.1
ollama==0.4.7
onnxruntime==1.20.1
openai==1.63.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
ply==3.11
portalocker==2.10.1
posthog==3.11.0
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.3
pyarrow==19.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.3
pymupdf4llm==0.0.17
pynndescent==0.5.13
pyparsing==3.2.1
pypdf==4.3.1
PyPDF2==3.0.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-Levenshtein==0.26.1
python-magic==0.4.27
python-multipart==0.0.18
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
qdrant-client==1.13.2
rank-bm25==0.2.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
requests-futures==1.0.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.11.4
sentence-transformers==3.4.1
sentencepiece==0.2.0
setuptools==75.8.0
shellingham==1.5.4
shortuuid==1.0.13
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
stdlibs==2024.12.3
streamlit==1.42.0
svgwrite==1.4.3
sympy==1.13.1
syncer==2.0.3
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
typer==0.15.1
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
uptrace==1.29.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==0.20.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==14.2
wn==0.9.5
wrapt==1.17.2
wsproto==1.2.0
xdg-base-dirs==6.0.2
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
zalgolib==0.2.2
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
